refactor(ui): extract worker lifecycle + media server from main.js (S5 Stage 1)
Pure move, no behavior change:

- ui/electron-main/worker.js — engine-worker lifecycle: fork of
  addon-worker.js, workerReady/addonError state, resolveSystemNodeExe,
  workerPathEnv, callWorker dispatch (30s timeout intact) and the
  SILENT_METHODS suppression set. Moved functions are byte-identical to
  the originals.
- ui/electron-main/media-server.js — local HTTP media server (Range
  support for <video>), bound port exposed via getMediaPort().
- main.js wires both via init({ log }) so all logging still lands in
  startup.log; outside reads of workerReady/addonError now go through
  isWorkerReady()/getAddonError()/setAddonError(), worker?.kill() ->
  killWorker(). Every IPC channel name, handler signature, timeout and
  gating check unchanged.
- electron-builder.json: add electron-main/**/* to the packaged-files
  whitelist (it is an explicit list; without this, packaged builds would
  miss the new modules).

WORLD-poll cadence and the shm video frame output are untouched
(AUDIT.md §4.3). Engine background tasks, autosave, persistence stores
and the handler tables stay in main.js for later stages.

Verified: npm run build && npm run test:smoke -> 3/3, both against the
working tree and against this exact committed main.js; vitest: 2478
passed, 2 failed = the pre-existing local-only mixer failures
(uncommitted FxBadge WIP), unrelated (no src/ changes here).
